addded keyboardInteractions file and finshed welcome.jsx
