+++ src/Welcome.css
@@ -2,9 +2,11 @@
 
 .welcome-container {
   position: absolute;
-  top: 3vh;
+  top: 5vh;
   left: 0;
+  z-index: 2;
   width: 100%;
+  height: 100%;
   display: flex;
   justify-content: space-between;
   padding: 3rem 2rem 3rem 2rem;
@@ -12,12 +14,13 @@
   
   letter-spacing: 0.1em;
   font-weight: 600;
-  background: none;
   color: black;
   pointer-events: none;
   z-index: 10;
   align-items: flex-start;
   box-sizing: border-box;
+  overflow: hidden;
+  overscroll-behavior: none;
 }
 
 .welcome-left {
@@ -47,4 +50,43 @@
 
 .label-block {
   white-space: nowrap;
+}
+
+.project-info {
+  position: absolute;
+  bottom: 3rem;
+  left: 2rem;
+  background-color: rgba(255, 255, 255, 0.85);
+  padding: 0.8rem;
+  max-width: 10rem;
+  font-size: 10px;
+  border-radius: 8px;
+  
+  z-index: 10;
+  pointer-events: auto;
+  border: 3px solid rgb(0, 0, 0);
+}
+
+.project-info img {
+  display: block;
+  max-width: 9rem;
+  max-height: 9rem;
+  height: auto;
+  border: 1px solid black;
+  border-radius: 6px;
+  margin-bottom: 0.6rem;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.see-more-link {
+  font-weight: 600;
+  text-decoration: none;
+  color: black;
+  display: inline-block;
+  margin-top: 4px;
+}
+
+.see-more-link:hover {
+  text-decoration: underline;
 }
